# dental-saas/backend/openapi/platform.yaml
openapi: 3.0.3
info:
  title: DentalSaaS Platform API
  version: 1.0.0
  description: |
    Sovereign Platform Plane API for DentalSaaS.
    Governs multi-tenant administration, capability enforcement
    and distributed governance (v19.3).
servers:
  - url: /api/platform
    description: Platform Plane (same-origin via Vite proxy / Nginx)

# ─────────────────────────────────────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# All paths are relative to /api/platform (server base URL)
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /core/users/org/{orgId}:
    get:
      summary: List org users (governance view)
      tags: [UserGovernance]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string } }
      responses:
        200:
          description: Org user list
  /core/users/org/{orgId}/{userId}:
    get:
      operationId: usersOrgUserDetail
      summary: Get org user detail (governance view)
      tags: [UserGovernance]
      parameters:
        - { name: orgId, in: path, required: true, schema: { type: string } }
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        200:
          description: User detail
...